requirements.txt
torch==2.6.0
torchvision>=0.10.0
transformers>=4.21.0
Pillow>=8.3.0
requests>=2.25.0
pandas>=1.3.0
tqdm>=4.62.0
numpy>=1.21.0
scikit-learn==1.6.1
matplotlib>=3.4.0
seaborn>=0.11.0
